backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==1.4.42
pymysql==1.1.0
databases==0.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic[email]==2.5.0
email-validator==2.0.0
pydantic-settings==2.1.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
openai==1.3.5
langchain==0.0.340
apscheduler==3.10.4
alembic==1.13.0
python-dotenv==1.0.0
loguru==0.7.2
prometheus-client==0.19.0